requirements.txt
channels
channels
codegen==1.0
coverage==4.5.1
Django==2.1.1
djangorestframework==3.8.2
patterns==0.3
pytz==2018.5
